updated tests for the spawnlist component
moved tests from pokemon detail to spawnlist component
